requirements.txt
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
blessed==1.21.0
causal_conv1d @ git+https://github.com/Dao-AILab/causal-conv1d.git@82867a9d2e6907cc0f637ac6aff318f696838548
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
contourpy==1.3.2
cycler==0.12.1
docker-pycreds==0.4.0
einops==0.8.1
filelock==3.18.0
fonttools==4.58.0
fsspec==2025.5.0
gitdb==4.0.12
GitPython==3.1.44
gpustat==1.1.1
huggingface-hub==0.31.4
idna==3.10
iniconfig==2.1.0
Jinja2==3.1.6
kiwisolver==1.4.8
mamba_ssm @ git+https://github.com/hustvl/Vim.git@dd0358ad1e42701f22afbefa0717cc8825cf9f45#subdirectory=mamba-1p1p1
MarkupSafe==3.0.2
matplotlib==3.10.3
mpmath==1.3.0
networkx==3.4.2
ninja==1.11.1.4
numpy==2.2.6
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-ml-py==12.575.51
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
omegaconf==2.3.0
packaging==25.0
pillow==11.2.1
platformdirs==4.3.8
pluggy==1.6.0
protobuf==6.31.0
psutil==7.0.0
pydantic==2.11.4
pydantic_core==2.33.2
pyparsing==3.2.3
pytest==8.3.5
python-dateutil==2.9.0.post0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
sentry-sdk==2.29.1
setproctitle==1.3.6
setuptools==80.8.0
six==1.17.0
smmap==5.0.2
sympy==1.14.0
timm==1.0.15
tokenizers==0.21.1
torch==2.7.0
torchaudio==2.7.0
torchinfo==1.8.0
torchvision==0.22.0
tqdm==4.67.1
transformers==4.52.3
triton==3.3.0
typing-inspection==0.4.1
typing_extensions==4.13.2
urllib3==2.4.0
wandb==0.19.11
warmup_scheduler @ git+https://github.com/ildoonet/pytorch-gradual-warmup-lr.git@7021d63a49106e22c79b40564a7d39930e7b0f53
wcwidth==0.2.13
wheel==0.45.1
